requirements.txt
flask==3.1.0
flask_sqlalchemy==3.1.1
flask_mysqldb==2.0.0
python-dotenv==1.0.1
jwt==1.3.1
flask==3.1.0
flask_sqlalchemy==3.1.1
flask_mysqldb==2.0.0
python-dotenv==1.0.1
PyJWT==2.6.0
flask-cors==4.0.0
pyopenssl==25.0.0